Adding option of submitting saved solutions
The option --submission_folder can be used to choose a location for saving solutions, this enables the student to keep several versions of its solutions.
If the --record_submission option is set the selected parts will be saved ini the folder given in --submission_folder option, if the option is missing the script will behave exactly as the previous version. If --record_submission is not passed and the --submission_folder is given, instead of computing the solutions the script will submit the solution previously saved int the selected folder.